// src/types/nav.type.ts
export type NavData =
	| {
			lgLogoImageURL: string;
			logoImageURL: string;
			mdLogoImageURL: string;
			smLogoImageURL: string;
			whiteTextLogoURL: string;
			smWhiteTextLogoURL: string;
			mdWhiteTextLogoURL: string;
			lgWhiteTextLogoURL: string;
			blackTextLogoURL: string;
			smBlackTextLogoURL: string;
			mdBlackTextLogoURL: string;
			lgBlackTextLogoURL: string;
	  }
	| undefined;